Add main entry with router and relocate pages
