+++ luna/src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import App from './App';
+import { BrowserRouter } from 'react-router-dom';
+import App from './pages/App.jsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
